+++ pmdarima/model_selection/_split.py
@@ -328,11 +328,20 @@ class SlidingWindowForecastCV(BaseTSCrossValidator):
             # TODO: what's a good sane default for this?
             window_size = max(3, n_samples // 5)
 
+        if window_size < 3:
+            raise ValueError("window_size must be > 2")
+
         indices = np.arange(n_samples)
-        for window_start in range(0, n_samples - h - window_size, step):
+        window_start = 0
+        while True:
             window_end = window_start + window_size
+            if window_end + h > n_samples:
+                break
+
             train_indices = indices[window_start: window_end]
             test_indices = indices[window_end: window_end + h]
+            window_start += step
+
             yield train_indices, test_indices