+++ pmdarima/model_selection/tests/test_split.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from pmdarima.compat.pytest import pytest_error_str
 from pmdarima.model_selection import RollingForecastCV, \
     SlidingWindowForecastCV, check_cv, train_test_split
 from pmdarima.datasets import load_wineind
@@ -94,3 +95,35 @@ def test_train_test_split():
     tr, te = train_test_split(y, test_size=10)
     assert te.shape[0] == 10
     assert_array_equal(y, np.concatenate([tr, te]))
+
+
+def test_bad_window_size():
+    cv = SlidingWindowForecastCV(window_size=2, step=1, h=4)
+    with pytest.raises(ValueError) as ve:
+        list(cv.split(y))
+    assert "> 2" in pytest_error_str(ve)
+
+
+def test_issue_364_bad_splits():
+    endog = y[:100]
+    cv = SlidingWindowForecastCV(window_size=90, step=1, h=4)
+    gen = cv.split(endog)
+
+    expected = [
+        (np.arange(0, 90), np.array([90, 91, 92, 93])),
+        (np.arange(1, 91), np.array([91, 92, 93, 94])),
+        (np.arange(2, 92), np.array([92, 93, 94, 95])),
+        (np.arange(3, 93), np.array([93, 94, 95, 96])),
+        (np.arange(4, 94), np.array([94, 95, 96, 97])),
+        (np.arange(5, 95), np.array([95, 96, 97, 98])),
+        (np.arange(6, 96), np.array([96, 97, 98, 99])),
+    ]
+
+    # should be 7
+    for i, (train, test) in enumerate(gen):
+        assert_array_equal(train, expected[i][0])
+        assert_array_equal(test, expected[i][1])
+
+    # assert no extra splits
+    with pytest.raises(StopIteration):
+        next(gen)